Change android package name from com.bloomreader to org.sil.bloom.reader
