+++ android/app/src/main/java/org/sil/bloom/reader/GetFromWifiPackage.java
@@ -1,4 +1,4 @@
-package com.bloomreader;
+package org.sil.bloom.reader;
 
 import com.facebook.react.ReactPackage;
 import com.facebook.react.bridge.NativeModule;
@@ -9,7 +9,13 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
-public class ShareApkPackage implements ReactPackage {
+/**
+    This boilerplate class is the required way of making native code accessible from JavaScript.
+    It exposes an instance of GetFromWifiModule so that it can be used from Javascript to fetch books from WiFi.
+    https://facebook.github.io/react-native/docs/native-modules-android#register-the-module
+ */
+
+public class GetFromWifiPackage implements ReactPackage {
     @Override
     public List<ViewManager> createViewManagers(ReactApplicationContext reactContext) {
         return Collections.emptyList();
@@ -18,7 +24,7 @@ public class ShareApkPackage implements ReactPackage {
     @Override
     public List<NativeModule> createNativeModules(ReactApplicationContext reactContext) {
         List<NativeModule> modules = new ArrayList<>();
-        modules.add(new ShareApkModule(reactContext));
+        modules.add(new GetFromWifiModule(reactContext));
         return modules;
     }
 }
