+++ android/app/src/main/java/org/sil/bloom/reader/MainActivity.java
@@ -1,4 +1,4 @@
-package com.bloomreader;
+package org.sil.bloom.reader;
 
 import android.net.Uri;
 import android.os.Bundle;
