+++ android/app/src/main/java/org/sil/bloom/reader/SyncServer.java
@@ -1,4 +1,4 @@
-package com.bloomreader;
+package org.sil.bloom.reader;
 
 import android.content.Context;
